flask-api/requirements.txt
Flask
python-dotenv
flask-cors
langchain-google-genai
pydantic
Pillow
pytesseract
